+++ CHANGES.rst
@@ -16,6 +16,8 @@ Bug Fixes
 - Fix 3D spectral extraction in cases where other types of data are loaded in addition
   to the cube. [#4181]
 
+- API hints toggle button to show in popout windows. [#4184]
+
 - Fix bug related to loading catalogs where units are determined from coordinate string. [#4199]
 
 - Fix aperture photometry plugin remaining visible in tray after all